// src/lib/types.ts
export interface ProjectDescription {
    title: string;
    description: string;
    images: string[];
    github: string;
    demo: string;
    longDescription: string;
    technologies: string[];
    features: string[];
}

export interface SkillDescription {
    category: string;
    items: string[];
}